Worked on errors, note to self: have const as a trailing return type instead of before the data type

+++ Peace Treaty V1.2/Operations/Source-Files/Units/Provinces/Buildings/BuildingsBASE.cpp
@@ -10,7 +10,7 @@ void BuildingsBASE::increaseLevel(int amount) {
 	level += amount;
 
 	for (int index = 0; index < 5; index++) { 
-		upgradeCosts.at(index) = level * baseUpgradeRates.at(index); 
+		upgradeCosts.at(index) = level * (int) baseUpgradeRates.at(index); 
 	}
 }
